# perf(es/minifier): Add some fast-path to the `MultiReplacer`