--- a/crates/swc_ecma_minifier/examples/minify-all.rs
+++ b/crates/swc_ecma_minifier/examples/minify-all.rs
@@ -23,8 +23,8 @@ fn main() {
 
     testing::run_test2(false, |cm, handler| {
         GLOBALS.with(|globals| {
-            files
-                .into_par_iter()
+            let _ = files
+                .into_iter()
                 .map(|path| -> Result<_> {
                     GLOBALS.set(globals, || {
                         let fm = cm.load_file(&path).expect("failed to load file");