--- a/crates/swc_ecma_minifier/scripts/instrument/all.sh
+++ b/crates/swc_ecma_minifier/scripts/instrument/all.sh
@@ -8,4 +8,4 @@
 
 
 export RUST_LOG=off
-cargo profile instruments -t time --example minify-all --release -- $@
+cargo profile instruments -t time --example minify-all --features swc_common/perf --release -- $@